Add pre-commit config and apply ruff lint fixes across codebase
- Add .pre-commit-config.yaml with ruff + ruff-format hooks (v0.8.6)
- Extend ruff ignore list: FBT001/002, SLF001, B905, RET504, PLW2901,
  PERF203, B023, SIM108, UP038, F841 for existing code patterns
- Fix RUF012: ClassVar annotation on CAMERA_NAME_MAP in flexiv env files
- Fix C408: dict() call -> literal in flexiv_rizon4_rt/real_env.py
- Fix RUF013: implicit Optional -> explicit | None in inspect_checkpoint.py
- Apply ruff auto-fixes across src/, packages/, scripts/, examples/

+++ scripts/train_pytorch.py
@@ -517,9 +517,9 @@ def train_loop(config: _config.TrainConfig):
                 break
 
             # The unified data loader returns (observation, actions) tuple
-            observation = jax.tree.map(lambda x: x.to(device), observation)  # noqa: PLW2901
-            actions = actions.to(torch.float32)  # noqa: PLW2901
-            actions = actions.to(device)  # noqa: PLW2901
+            observation = jax.tree.map(lambda x: x.to(device), observation)
+            actions = actions.to(torch.float32)
+            actions = actions.to(device)
 
             # Update LR
             for pg in optim.param_groups:
